Minichain_v0/requirements.txt
# MiniChain v0 Dependencies
# Cryptography: Ed25519 signatures
pynacl>=1.5.0
